Move rpc config structs
